OneDrive/Documents/SuperAnnuation/Sayli-dh/superannuation-ai-assistant/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Limiter==3.5.0
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
openai==0.28.0
requests==2.31.0
redis==4.6.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
matplotlib==3.7.2
seaborn==0.12.2
yfinance==0.2.21
alpha-vantage==2.3.1
schedule==1.2.0
APScheduler==3.10.4
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
celery==5.3.6
werkzeug==2.3.7
marshmallow==3.20.1
flasgger==0.9.7.1
python-dateutil==2.8.2
openpyxl==3.1.2
xlrd==2.0.1